librabbitmq: protect the password from being interpreted as flags
If the password starts with a dash (`-`), it would otherwise be
interpreted as a series of single-character flags by rabbitmqctl,
making the password change fail.

This can be hard to detect because the container will immediately
be restarted and the password change is not retried by the scripts
because the user exists already. So unless you happen to be
watching the logs from the beginning, it can be a very confusing
situation.