Introduce flag files to mark complete initialization
Without these, it is possible that the first attempt at
initializing RabbitMQ fails spuriously (for example, due to a
timeout while starting the RabbitMQ service in background because
the cluster is overloaded).

In that case, the data directory may still be non-empty afterward,
causing the subsequent run of the container to assume that
configuration has completed. If this is used with the secure way
of setting the user password, it will cause the user to not be
configured at all, making subsequent attempts to interact with
the cluster fail.

To avoid this, the data directory is now marked with two different
files depending on the phase of initialization:

- If the directory is empty, it is marked with an "init" file to
  note that initialization needs to happen.
- If the directory is non-empty and marked with an "init" file,
  but no "complete" file, re-try the initialization procedure.
- If the directory is non-empty and contains the "complete"
  marker, proceed with normal startup.
- If the directory is non-empty, but no marker file is present,
  we assume that this is an upgrade with an existing, healthy
  cluster and mark the initialization as complete by placing
  both markers.